separated client logic from review/results pages | updated api calls in all review/result components to use base_url/not be fucking stupid | removed utils and type files
